Add UpdateIssuesJob
We will now be using the GitHub API for 2 things:

- finding new issues to create (CreateIssuesJob). This only looks for
open issues. It update attributes other than the "open/closed" status
if the API call returns Issues we have already created, or create new
Issues if they are not already in the database.
- updating existing issues (UpdateIssuesJob). This looks for "open" and
"closed" issues, so that we can stop showing Issues that are no longer
available. It also updates other attributes.

From what I could see there is no way to make 1 API call to get
updated information just concerning the Issues that are already in the
database. So the solution here would be to check for all recent issues
"open" or "closed". We could run the job to create 40 issues, and the day
after, the job to update 100 issues to be sure we get data from issues
that we've created a while ago. From what I could see there is generally
not a lot of issues created/updated/closed everyday on GitHub, so that
should hopefully ensure we keep track of closed issues. We can give it a
try like this and see if it needs tweaking/how regularly we should run
the jobs.