Fix li elements
The default list style for Tailwind is list-style-type: none.
This is why we couldn't see bullet points or numbers when using li
elements in unordered or ordered lists. So I have reverted this as per
the answer of this StackOverflow question:
https://stackoverflow.com/questions/69276276/why-tailwind-list-style-type-is-not-working

This is the base style for Preflight:
https://tailwindcss.com/docs/preflight#lists-are-unstyled

And these are the possible classes for list style types in Tailwind:
https://tailwindcss.com/docs/list-style-type