Change style of container with link to guide
I implemented Claire's suggestions for this container.
There is still a "container-1" used on the dashboard but I think we'll
need to see what we want to do about it in a future PR, so for now I have
just created a different class "guide-container" for the container that
links to the beginner's guide.