Change "users" to move to logging in with GitHub
We will only be using GitHub to authenticate, so we no longer need the
fields related to email and passwords.

The info we get from GitHub does not always include a password if users
chose to keep it private, so we would have to make the presence of an
email conditional to the mean of authentication if we wanted to keep
both the authentication by email and by GitHub. But after discussing this
on Slack, we decided to only support authentication with GitHub, as this
app mostly makes sense if users have a GitHub account.

I've added an attribute to get the url of the GitHub avatar so we can use it
when people create or comment on a "call to action".