Support authentication with GitHub omniauth
Since we are no longer supporting authentication by email and password,
I had to remove most of the default Devise modules.

I have removed the "format" on the validation for gh_url because it did
not seem correct from what I tested, but I don't know what all the
possibilities could be. We will get this info from GitHub anyway so it
should be reliable.

We no longer need a button to sign up, people can just use 1 button to
sign up or log in with GitHub.